# Conflicts:
#	package-lock.json
#	package.json
#	src/web/assets/cp/dist/css/craft.css.map
#	src/web/assets/cp/dist/js/Craft.min.js
#	src/web/assets/cp/dist/js/Craft.min.js.map

+++ src/controllers/UsersController.php
@@ -1212,18 +1212,27 @@ class UsersController extends Controller
         $user->firstName = $this->request->getBodyParam('firstName', $user->firstName);
         $user->lastName = $this->request->getBodyParam('lastName', $user->lastName);
 
-        // There are some things only admins can change
-        if ($currentUser && $currentUser->admin) {
+        // New users should always be initially saved in a pending state,
+        // even if an admin is doing this and opted to not send the verification email
+        if ($isNewUser) {
+            $user->pending = true;
+        }
+
+        if ($canAdministrateUsers) {
             $user->passwordResetRequired = (bool)$this->request->getBodyParam('passwordResetRequired', $user->passwordResetRequired);
+        }
 
-            // Is their admin status changing?
-            if (($adminParam = $this->request->getBodyParam('admin', $user->admin)) != $user->admin) {
-                // Making someone an admin requires an elevated session
-                if ($adminParam) {
-                    $this->requireElevatedSession();
-                }
-
-                $user->admin = (bool)$adminParam;
+        // Is their admin status changing?
+        if (
+            $currentUser &&
+            $currentUser->admin &&
+            ($adminParam = $this->request->getBodyParam('admin', $user->admin)) != $user->admin
+        ) {
+            if ($adminParam) {
+                $this->requireElevatedSession();
+                $user->admin = true;
+            } else {
+                $user->admin = false;
             }
         }
 
@@ -1447,7 +1456,7 @@ class UsersController extends Controller
                 'html' => $this->_renderPhotoTemplate($user),
             ]);
         } catch (Throwable $exception) {
-            if (isset($fileLocation)) {
+            if (isset($fileLocation) && file_exists($fileLocation)) {
                 FileHelper::unlink($fileLocation);
             }
 
@@ -1626,7 +1635,7 @@ class UsersController extends Controller
             $entryCount = Entry::find()
                 ->sectionId($section->id)
                 ->authorId($userIds)
-                ->siteId('*')
+                ->site('*')
                 ->unique()
                 ->status(null)
                 ->count();
